tests/nc-client: clarify client network errors
Changing warn0() to warnp() revealed that when the Travis-CI tests fail, it's
due to a broken pipe.  Definitely a good thing to know!

It's true that given the specific values inside this file's send_input()
function, we're not expecting to see a lenwrit which is neither -1 nor
send->nchars.  However, I think the code is more clear this way, and it's
entirely probable that I'll end up copying bits of this file the next time I
want to send data over a network.  Also, if we did somehow see an unexpected
value for lenwrit, I think it would be nice to have a specific error message
for that case.